Update docs, remove auto state getter memoization
